requirements.txt
# Django
Django==1.10.7

# Environmental var configuration
django-configurations==2.0

# For the persistence stores
psycopg2==2.7.2
dj-database-url==0.4.2

# Models
django-model-utils==3.0.0

# Rest apis
djangorestframework==3.6.3
Markdown==2.6.8
django-filter==1.0.4

# Static files
whitenoise==3.3.0

# Backend pool
ramos==1.0.0

# Media files
boto==2.48.0
django_unique_upload==0.2.1

# Image manipulation
django-versatileimagefield==1.7.1

# For asynchronous queuing
django-rq==0.9.6

# Time zones support
pytz==2017.2


# WSGI Handler
gevent==1.2.2
gunicorn==19.7.1

# Static and Media Storage
django-storages-redux==1.3.3


# Mocking
mock==2.0.0

# Fixtures
factory-boy==2.8.1

# Test Runner
django-nose==1.4.4
nose-progressive==1.5.1

# Code Coverage
coverage==4.4.1

# Style Enforcement
flake8==3.3.0

mysqlclient==1.3.12
